Move provider to factory
