+++ composer.json
@@ -20,7 +20,6 @@
     "require": {
         "php": "^7.4|^8.0",
         "psr/container": "^1.0",
-        "yiisoft/di": "^3.0@dev",
         "yiisoft/event-dispatcher": "^3.0@dev",
         "yiisoft/friendly-exception": "^1.0",
         "yiisoft/injector": "^1.0"
